requirements.txt
selenium
python-dotenv
azure-storage-blob